Add missing manifest attributes to Fabric Transformers
